release: Add support for creating ZFS-based images
The change extends vmimage.subr to handle a new parameter, VMFS, which
should be equal to either "ufs" or "zfs".  In the latter case, rather
than using makefs(8) to create the filesystem image, we use md(4) to
create a zpool with a single disk vdev, mounted at $DESTDIR, and install
the FreeBSD distribution there.  The zpool has the autoexpand property
set so that growfs works as with UFS.

This has a couple of caveats.  First, ZFS images cannot be built in a
jail, as the kernel forbids creation of zpools when the creating process
is jailed.  The pool will also contain random GUIDs for the pool and
vdev, so images are not completely reproducible.  I'm working on
extending makefs(8) to create the filesystem image using libzpool, which
if successful will fix both problems, but in the meantime this solution
works well and provides an often-requested feature without much added
complexity.

The dataset layout is the same as that used by bsdinstall and "poudriere
image".  There is a distinct lack of error handling in the release build
scripts, so I haven't tried to handle errors in my extensions, but that
would be a good next step IMO.