bhyve: Address an unused parameter warning in the smbios code
The compiler was warning that the "size" parameter to
smbios_generic_initializer() was unused.  This parameter is apparently
used to populate the "maximum structure size" field in the SMBIOS entry
point, but we are always setting it to zero, apparently without ill
effect.  The SMBIOS specification doesn't really describe how it is to
be used, and FreeBSD and Linux appear to ignore it.  So just remove all
of the code used to initialize this field.

No functional change intended.