dtrace: Implement a "regs" variable
This allows invop-based providers (i.e., fbt and kinst) to expose the
register file of the CPU at the point where the probe fired.  It does
not work for SDT providers because they're implemented as plain function
calls and so don't get a saved register file.  It's not clear what
semantics regs should have for them anyway.

This is similar to "uregs", which nominally provides access to the
userspace registers.  As a side note, though, it doesn't actually do
this in all cases, since td_frame is set any time a thread enters an
exception or interrupt handler.  For example, if a thread takes an
interrupt or page fault while in kernel mode, "uregs" will provide
kernel registers, not user mode registers.  On x86, the flags register
can at least tell you which privilege mode the CPU was in when the
registers were saved, so it's possible to check from D.