inpcb: Upgrade some lock assertions
INP_HASH_LOCK_ASSERT() asserts that the caller is either in an SMR read
section, or holds the hash lock, which is not in fact a rwlock but is
simply a mutex.  However, most users of INP_HASH_LOCK_ASSERT() really
require the mutex, so let's tighten the assertions a bit.