arm64: Disable PAC when booting on a Cortex-A78C
There is an outstanding erratum documented in SDEN-1707916 for this
core, stating that PAC is effectively broken (PAC instructions may
exhibit unexpected behaviours).  This is similar but not identical to a
problem I see on the MS Devkit wherein the boot CPU hangs when
ptrauth_thread0() loads the APIA key.

Disable PAC if the boot CPU is a ARM Cortex-A78C.  This is enough to
allow the kernel to boot.  However, it is not general as it does not
handle the situation where the kernel boots on an unaffected CPU but the
system contains a Cortex-A78C.