netinet tests: Enable parallelization of forwarding tests
Each of these tests creates a vnet jail and an epair, and assigns an IP
address to the host side of the epair.  They cannot be run in
parallel due to IP address conflicts.

Enable parallelization by creating a second jail per test and putting
the other end of the epair in the other jail.  This does not make the
tests significantly more complicated.

We have other tests which have the same problem.  I'm not sure whether
such a solution is appropriate there since it would involve a lot of
churn.  This revision is meant to garner feedback and see what folks
think.