buf: Preallocate KVA for paging bufs
A TSLOG flame graph of a firecracker microVM boot shows that we spend a
non-negligible amount of time allocating KVA pbuf-bu-pbuf.  Try to speed
it up a bit by attempting to allocate all of the required KVA up front
and using a bump allocator in pbuf_init().

This also means that we no longer allocate boundary tags for each pbuf,
and that isn't necessary anyway since pbuf KVA ranges are never freed
back to the VM.

Reported by:	cperciva