vtnet: Ensure that L3 protocol headers are aligned when necessary
This addresses alignment faults which occur when running an armv7 kernel
in QEMU: the L3 protocol implementations effectively assume that headers
are aligned.

Modify vtnet's buffer allocation routine to add a 2-byte offset to the
beginning of the data buffer.  Add 2 to the interface MTU so that we
allocate sufficiently large clusters.

PR:		271288
Sponsored by:	Klara, Inc.
Sponsored by:	Stormshield