build: Introduce MK_MSAN
As with MK_ASAN and MK_UBSAN, this lets one build parts of the tree, or
the whole tree, with LLVM's Memory Sanitizer enabled.  This enables
runtime detection of uses of uninitialized memory.

MSAN has some extra constraints relative to ASAN and UBSAN:
- All code loaded into a process must be instrumented, otherwise false
  positives can arise easily.
- An MSAN-instrumented DSO cannot be linked into an uninstrumented
  executable.
- MSAN-instrumented executables must be position-independent.  In
  particular, anything compiled with MK_PIE=no cannot be instrumented.

In principle, libc.so does not need to be instrumented because the MSAN
runtime provides interceptors which handle shadow map updates for
various commit libc functions.  However:
- Some commonly used libc symbols are not intercepted currently (e.g.,
  getc() and cgetent(), the latter is used by most ncurses
  applications).
- There is at least one case where libc itself will generate a false
  positive: fts_sort() allocates memory using realloc(), which is
  intercepted, initializes the array inline, then sorts it with qsort(),
  which is also intercepted.  MSAN ends up reporting that the memory is
  uninitialized.

This commits introduces the glue needed to compile parts of the base
system with MSAN enabled.  It's not yet possible to run the entire base
system with MSAN enabled, issues with instrumentation of
libc/libthr/libsys need to be resolved first.  However, it's enough for
some purposes.