sdt: Prototype implementation of SDT probes using hot-patching
This isn't functional but simply serves to evaluate the performance
overhead of an asm-goto-based approach.

The idea here is to avoid a memory access and conditional branch.
Instead, the probe is represented by an "unreachable" unconditional
function call.  asm goto is used to store the address of the probe site
(represented by a no-op sled) and the address of the function call.

When the probe is enabled, the no-op sled is overwritten with a jmp to
the label.

I verified that llvm 17 moves argument marshalling code and the
sdt_probe() function call out-of-line, i.e., to the end of the function.