libc/tests: Fix installation without MK_TOOLCHAIN
There is some strange conditional logic here to avoid building a
particular test if a certain UBSAN library isn't present.  This causes
build failures for me when doing an "installworld WITHOUT_TOOLCHAIN=",
which I do frequently during tests.  (This failure appeared for me only
very recently, I'm not quite sure why.)

I presume the problem is that SYSROOT is unset during installworld, so
the build sees the host's copy of libclang_rt.ubsan_standalone.a and
then tries to install a binary that wasn't built during buildworld.
Assuming that this is correct, try to make the check a bit less fragile.